requirements.txt.txt
python-magichue==0.2.9.1